+++ package/lib/src/addon/adaptive_padding.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:exprollable_page_view/src/core/controller.dart';
-import 'package:exprollable_page_view/src/core/view.dart';
+import 'package:exprollable_page_view/src/core/view/view.dart';
 import 'package:exprollable_page_view/src/internal/utils.dart';
 import 'package:flutter/widgets.dart' hide Viewport;
