+++ package/lib/src/addon/modal.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 import 'dart:ui';
 
 import 'package:exprollable_page_view/src/core/controller.dart';
-import 'package:exprollable_page_view/src/core/view.dart';
+import 'package:exprollable_page_view/src/core/view/view.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart' hide Viewport;
