+++ package/lib/src/core/controller.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 
 import 'package:exprollable_page_view/src/internal/scroll.dart';
 import 'package:exprollable_page_view/src/internal/utils.dart';
-import 'package:exprollable_page_view/src/core/view.dart';
+import 'package:exprollable_page_view/src/core/view/view.dart';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart' hide Viewport;
