+++ package/lib/src/core/core.dart
@@ -1,4 +1,4 @@
 library core;
 
-export 'package:exprollable_page_view/src/core/view.dart';
+export 'package:exprollable_page_view/src/core/view/view.dart';
 export 'package:exprollable_page_view/src/core/controller.dart';
