+++ package/lib/src/core/view/view.dart
@@ -1,0 +1,3 @@
+export 'package:exprollable_page_view/src/core/view/exprollable_page_view.dart';
+export 'package:exprollable_page_view/src/core/view/page_configuration.dart'
+    hide InheritedPageController;
